requirements.txt
django
gunicorn
django-heroku
requests
python-decouple
pandas
djangorestframework
djangorestframework-jwt
djangorestframework-simplejwt